test: change import order
